FIX: remove package  even if it has a version specifier
